requirements.txt
tensorflow>=2.15.0
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
streamlit>=1.30.0
Pillow>=10.0.0
